requirements.txt
yt-dlp 
webvtt-py
pymongo
python-dotenv